# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SoundTrip Backend",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/songs": {
      "post": {
        "tags": [
          "songs"
        ],
        "summary": "Create Song",
        "operationId": "create_song_api_v1_songs_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SongCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SongRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "songs"
        ],
        "summary": "List Songs",
        "operationId": "list_songs_api_v1_songs_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SongRead"
                  },
                  "title": "Response List Songs Api V1 Songs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/songs/{song_id}": {
      "get": {
        "tags": [
          "songs"
        ],
        "summary": "Get Song",
        "operationId": "get_song_api_v1_songs__song_id__get",
        "parameters": [
          {
            "name": "song_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Song Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SongRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "songs"
        ],
        "summary": "Update Song",
        "operationId": "update_song_api_v1_songs__song_id__put",
        "parameters": [
          {
            "name": "song_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Song Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SongUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SongRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "songs"
        ],
        "summary": "Delete Song",
        "operationId": "delete_song_api_v1_songs__song_id__delete",
        "parameters": [
          {
            "name": "song_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Song Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/influences": {
      "post": {
        "tags": [
          "influences"
        ],
        "summary": "Create Influence",
        "operationId": "create_influence_api_v1_influences_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InfluenceCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InfluenceRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "influences"
        ],
        "summary": "List Influences",
        "operationId": "list_influences_api_v1_influences_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InfluenceRead"
                  },
                  "title": "Response List Influences Api V1 Influences Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/influences/{influence_id}": {
      "get": {
        "tags": [
          "influences"
        ],
        "summary": "Get Influence",
        "operationId": "get_influence_api_v1_influences__influence_id__get",
        "parameters": [
          {
            "name": "influence_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Influence Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InfluenceRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "influences"
        ],
        "summary": "Update Influence",
        "operationId": "update_influence_api_v1_influences__influence_id__put",
        "parameters": [
          {
            "name": "influence_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Influence Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InfluenceUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InfluenceRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "influences"
        ],
        "summary": "Delete Influence",
        "operationId": "delete_influence_api_v1_influences__influence_id__delete",
        "parameters": [
          {
            "name": "influence_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Influence Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "InfluenceCreate": {
        "properties": {
          "category_id": {
            "type": "integer",
            "title": "Category Id"
          },
          "key": {
            "type": "string",
            "maxLength": 160,
            "minLength": 1,
            "title": "Key"
          },
          "label": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Label"
          }
        },
        "type": "object",
        "required": [
          "category_id",
          "key",
          "label"
        ],
        "title": "InfluenceCreate"
      },
      "InfluenceRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "category_id": {
            "type": "integer",
            "title": "Category Id"
          },
          "key": {
            "type": "string",
            "title": "Key"
          },
          "label": {
            "type": "string",
            "title": "Label"
          }
        },
        "type": "object",
        "required": [
          "id",
          "category_id",
          "key",
          "label"
        ],
        "title": "InfluenceRead"
      },
      "InfluenceUpdate": {
        "properties": {
          "category_id": {
            "type": "integer",
            "title": "Category Id"
          },
          "key": {
            "type": "string",
            "maxLength": 160,
            "minLength": 1,
            "title": "Key"
          },
          "label": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Label"
          }
        },
        "type": "object",
        "required": [
          "category_id",
          "key",
          "label"
        ],
        "title": "InfluenceUpdate"
      },
      "SongCreate": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Title"
          },
          "artist": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Artist"
          },
          "album": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Album"
          },
          "release_year": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 3000.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Release Year"
          },
          "duration_seconds": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Seconds"
          },
          "language": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 80
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "time_period_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time Period Id"
          },
          "primary_geography_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Primary Geography Id"
          }
        },
        "type": "object",
        "required": [
          "title",
          "artist"
        ],
        "title": "SongCreate"
      },
      "SongRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "artist": {
            "type": "string",
            "title": "Artist"
          },
          "album": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Album"
          },
          "release_year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Release Year"
          },
          "duration_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Seconds"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "time_period_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time Period Id"
          },
          "primary_geography_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Primary Geography Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "artist",
          "album",
          "release_year",
          "duration_seconds",
          "language",
          "notes",
          "time_period_id",
          "primary_geography_id",
          "created_at"
        ],
        "title": "SongRead"
      },
      "SongUpdate": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Title"
          },
          "artist": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Artist"
          },
          "album": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Album"
          },
          "release_year": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 3000.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Release Year"
          },
          "duration_seconds": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Seconds"
          },
          "language": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 80
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "time_period_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time Period Id"
          },
          "primary_geography_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Primary Geography Id"
          }
        },
        "type": "object",
        "required": [
          "title",
          "artist"
        ],
        "title": "SongUpdate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}